Fix all size_t to float conversion warnings for strict GCC compliance
- Add explicit static_cast<float>() for all size_t arithmetic operations
- Fix patterns: i multiplication, (i+1) expressions, and i/fft_bins divisions
- Ensures clean build on Linux with strict conversion warnings
- All heap tests continue to pass (1155 assertions)

+++ tests/test_gradients.cpp
@@ -217,7 +217,7 @@ TEST_CASE("Frequency Gradient Calculations", "[gradients][frequency]")
         // Generate sinusoidal phase pattern
         for (size_t i = 0; i < fft_bins; ++i)
         {
-            phases[i] = 0.5f * std::sin(2.0f * 3.14159f * i / fft_bins);
+            phases[i] = 0.5f * std::sin(2.0f * 3.14159f * static_cast<float>(i) / static_cast<float>(fft_bins));
         }
 
         REQUIRE(rtpghi::calculate_freq_gradients(phases.data(), fft_bins, freq_step,
